+++ EveUtils.Migrations.Client.Sqlite/Migrations/20260912080626_AddActivitySummaryStartedAtUtcIndex.cs
@@ -1,0 +1,27 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace EveUtils.Migrations.Client.Sqlite.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddActivitySummaryStartedAtUtcIndex : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateIndex(
+                name: "IX_ActivitySummary_StartedAtUtc",
+                table: "ActivitySummary",
+                column: "StartedAtUtc");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropIndex(
+                name: "IX_ActivitySummary_StartedAtUtc",
+                table: "ActivitySummary");
+        }
+    }
+}